fix: rename middleware.ts to proxy.ts (Next.js 16 convention)
